requirements.txt
# Core dependencies
phidata>=1.0.0
crewai>=0.1.0
langchain>=0.1.0
openai>=1.0.0
duckduckgo-search>=3.0.0

# ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0

# Web and API
fastapi>=0.70.0
streamlit>=1.0.0
uvicorn>=0.15.0

# Voice Processing
speechrecognition>=3.8.1
pyttsx3>=2.90

# Visualization
matplotlib>=3.4.0
plotly>=5.3.0

# Development
pytest>=6.0.0
black>=21.0
isort>=5.0.0
mypy>=0.900

# Additional Tools
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0 